Prettier plugin

+++ contracts/Migrations.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.8.4;
 
 /**
-  * Migrations
-  * A helper contract to manage truffle migrations
-**/
+ * Migrations
+ * A helper contract to manage truffle migrations
+ **/
 contract Migrations {
   address public owner;
-  uint public last_completed_migration;
+  uint256 public last_completed_migration;
 
   constructor() {
     owner = msg.sender;
@@ -17,7 +17,7 @@ contract Migrations {
     if (msg.sender == owner) _;
   }
 
-  function setCompleted(uint completed) public restricted {
+  function setCompleted(uint256 completed) public restricted {
     last_completed_migration = completed;
   }
 }
